Fix project/solution references on non-windows platforms
We had a windows-only "dotnet.exe" command. This commit changes it so we
call either "dotnet.exe" or "dotnet" depending on the platform.

Additionally, this was covered by an integration test that was failing
on Ubuntu, so this PR additionally enables Ubuntu in the CI tests.

Just for good measure, I also added a `.vscode` directory for easier
development on Ubuntu. 🚀