Always respect user's completion keybinding
If the user has a configured completion keybinding (or is using the
default keybinding, e.g. Tab), make sure that we always use that
keybinding as a completion trigger.

For Tab specifically, it appears to usually be returned by Roslyn, but
in some cases (e.g. typing using directives) it does not.