Revert 7f2c801 as it interferes with usability
Need to figure out what's wrong with our roslyn completion commit
characters. Reverting this as it's not a good direction for a fix.

There's a unit test commented out
(ObjectInitialization_CompletionDoesNotInterfere) for now, as we're
working on better completion commit character usage this could serve as
a useful test case.