Simplify disassembly output for simple cases.
In the case where a simple statement like "var sum = 5 + 3;" is submitted for decompilation, we send back a ton of text: the <Module> declaration, the generated Program, its constructor, as well as the generated `<Main>$` method.

This PR simplifies the output so that in cases where there's only a Program.Main, we return just the contents of the Program.Main method. In other cases, like where functions and types are created, we still return the full output.

This PR also unfortunately removes the syntax highlighting attempt for IL -- we were co-opting the roslyn c# highlighting logic (which happened to mostly work, but now throws an exception).